requirements.txt
diffusers
matplotlib
Pillow
torch>=1.9.0
torchvision>=0.10.0
numpy
streamlit